Trait zh_open_data::loader::Loader

trait Loader<K, V> {
    fn process_line(map: &mut HashMap<K, V>, line: String);

    fn load_into_map(&self, map: &mut HashMap<K, V>, paths: &[String]) { ... }

    fn get_map(&self, paths: &[String]) -> Result<HashMap<K, V>> { ... }

    fn load_into(map: &mut HashMap<K, V>, path: &str) -> Result<()> { ... }
}

Required methods

fn process_line(map: &mut HashMap<K, V>, line: String)

Given a map with concrete types, and a line, do something : process the line and put the result into the map

Provided methods

fn load_into_map(&self, map: &mut HashMap<K, V>, paths: &[String])

Load and map given paths into an existing map

fn get_map(&self, paths: &[String]) -> Result<HashMap<K, V>>

Get a single map with given paths loaded and mapped into it

fn load_into(map: &mut HashMap<K, V>, path: &str) -> Result<()>

Load a given file into a given map